Update Nuget packages in VS extension
