+++ source/VisualStudio.Extension/ProjectProperties.cs
@@ -1,0 +1,48 @@
+namespace nanoFramework.Tools.VisualStudio.Extension
+{
+    using System;
+    using System.Collections.Generic;
+    using System.ComponentModel.Composition;
+    using System.Linq;
+    using System.Text;
+    using System.Threading.Tasks;
+    using Microsoft.VisualStudio.ProjectSystem;
+    using Microsoft.VisualStudio.ProjectSystem.Properties;
+
+    [Export]
+    internal partial class ProjectProperties : StronglyTypedPropertyAccess
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ProjectProperties"/> class.
+        /// </summary>
+        [ImportingConstructor]
+        public ProjectProperties(ConfiguredProject configuredProject)
+            : base(configuredProject)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ProjectProperties"/> class.
+        /// </summary>
+        public ProjectProperties(ConfiguredProject configuredProject, string file, string itemType, string itemName)
+            : base(configuredProject, file, itemType, itemName)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ProjectProperties"/> class.
+        /// </summary>
+        public ProjectProperties(ConfiguredProject configuredProject, IProjectPropertiesContext projectPropertiesContext)
+            : base(configuredProject, projectPropertiesContext)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ProjectProperties"/> class.
+        /// </summary>
+        public ProjectProperties(ConfiguredProject configuredProject, UnconfiguredProject unconfiguredProject)
+            : base(configuredProject, unconfiguredProject)
+        {
+        }
+    }
+}
