+++ source/VisualStudio.Extension/ProjectSystem/ProjectProperties.cs
@@ -14,43 +14,9 @@ namespace nanoFramework.Tools.VisualStudio.Extension
     [ExcludeFromCodeCoverage]
     internal partial class ProjectProperties : StronglyTypedPropertyAccess
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ProjectProperties"/> class.
-        /// </summary>
-        [ImportingConstructor]
-        public ProjectProperties([Import] ConfiguredProject configuredProject)
-            : base(configuredProject)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ProjectProperties"/> class.
-        /// </summary>
-        public ProjectProperties(ConfiguredProject configuredProject, string file, string itemType, string itemName)
-            : base(configuredProject, file, itemType, itemName)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ProjectProperties"/> class.
-        /// </summary>
-        public ProjectProperties(ConfiguredProject configuredProject, IProjectPropertiesContext projectPropertiesContext)
-            : base(configuredProject, projectPropertiesContext)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ProjectProperties"/> class.
-        /// </summary>
-        public ProjectProperties(ConfiguredProject configuredProject, UnconfiguredProject unconfiguredProject)
-            : base(configuredProject, unconfiguredProject)
-        {
-        }
-
         public new ConfiguredProject ConfiguredProject
         {
             get { return base.ConfiguredProject; }
         }
-
     }
 }
